+++ src/Game/Werewolf/Messages.hs
@@ -59,7 +59,7 @@ module Game.Werewolf.Messages (
     scapegoatLynchedMessage, villageIdiotLynchedMessage,
 
     -- ** Error messages
-    playerHasAlreadyVotedMessage, playerCannotLynchVillageIdiotMessage,
+    playerHasAlreadyVotedMessage,
 
     -- * Werewolves' turn messages
     playerMadeDevourVoteMessage, playerDevouredMessage, noPlayerDevouredMessage,
@@ -465,16 +465,13 @@ scapegoatLynchedMessage name = publicMessage $ T.unwords
 
 villageIdiotLynchedMessage :: Text -> Message
 villageIdiotLynchedMessage name = publicMessage $ T.concat
-    [ "Just as the townsfolk tie", name, "up to the pyre, a voice in the crowd yells out."
+    [ "Just as the townsfolk tie ", name, " up to the pyre, a voice in the crowd yells out."
     , " \"We can't burn ", name, "! He's that oaf, you know, John's boy!\""
     , " The Village Idiot is quickly untied and apologised to."
     ]
 
 playerHasAlreadyVotedMessage :: Text -> Message
 playerHasAlreadyVotedMessage to = privateMessage to "You've already voted!"
-
-playerCannotLynchVillageIdiotMessage :: Text -> Message
-playerCannotLynchVillageIdiotMessage to = privateMessage to "You cannot lynch the Village Idiot!"
 
 playerMadeDevourVoteMessage :: Text -> Text -> Text -> Message
 playerMadeDevourVoteMessage to voterName targetName = privateMessage to $ T.concat