+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Revisions*
 
 * Fixed bug where the Wolf-hound's turn messages would be displayed on every round. ([#87](https://github.com/hjwylde/werewolf/issues/87))
+* Fixed a bug causing the Angel's joining Villagers message to be shown every round. ([#95](https://github.com/hjwylde/werewolf/issues/95))
 
 #### v0.4.7.0