+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Fixed bug where the Wolf-hound's turn messages would be displayed on every round. ([#87](https://github.com/hjwylde/werewolf/issues/87))
 * Fixed a bug causing the Angel's joining Villagers message to be shown every round. ([#95](https://github.com/hjwylde/werewolf/issues/95))
+* Fixed Village Idiot text to have spaces around the name. ([#87](https://github.com/hjwylde/werewolf/issues/87))
+* Fixed a bug where the Werewolves couldn't win if it was down to 1 Werewolf and the Village Idiot. ([#88](https://github.com/hjwylde/werewolf/issues/88))
 
 #### v0.4.7.0
 
@@ -19,6 +21,13 @@
 * Fixed the grammar on the first Werewolves' turn messages.
 * Moved Wolf-hound's turn to before the Seer's so that the Seer may see his allegiance properly.
 * Restricted specifying `Simple Villager` or `Simple Werewolf` as extra roles.
+
+#### v0.4.6.1
+
+*Revisions*
+
+* Fixed Village Idiot text to have spaces around the name. ([#87](https://github.com/hjwylde/werewolf/issues/87))
+* Fixed a bug where the Werewolves couldn't win if it was down to 1 Werewolf and the Village Idiot. ([#88](https://github.com/hjwylde/werewolf/issues/88))
 
 #### v0.4.6.0